requirements.txt
django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
psycopg2
